backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
sqlalchemy
asyncpg
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
boto3
openai
redis
pyjwt